requirements.txt
flask==2.2.2
pandas==1.5.3
numpy==1.23.5
scikit-learn==1.1.3
shap==0.41.0
matplotlib==3.6.2
seaborn==0.12.1
openai==0.27.0
python-dotenv